Mac: read info on mounts not backed by phys. devices
Example: a SMB mount
